useContext works for main header
