+++ react-front-end/src/components/States/UIStateProvider.js
@@ -5,8 +5,16 @@ export default function UIStateProvider(props) {
   const [uiState, setUIState] = useState({
     left: false,//is container open or closed
     right: false,
-    disabled: true //disable FAB icons and side containers when header is expanded
-  })
+    disabled: true, //disable FAB icons and side containers when header is expanded
+    chooseCountry: true,
+    showTrends: false,
+    currentCountry: '',
+    currentTrend: '',
+  });
+
+  // const [headerState, setHeaderState] = useState({
+  // })
+
 
   // TODO - AR ask about diff between prev or just spreading state
   const toggleLeft = () => {
@@ -22,12 +30,51 @@ export default function UIStateProvider(props) {
     setUIState(prev => ({ ...prev, disabled: true }))
   };
 
+  const toggleChooseCanada = () => {
+    setUIState(prev => ({
+      ...prev,
+      chooseCountry: !uiState.chooseCountry,
+      showTrends: !uiState.showTrends,
+      currentCountry: 'canada'
+    }));
+  }
+  const toggleChooseUsa = () => {
+    setUIState(prev => ({ 
+      ...prev,
+      chooseCountry: !uiState.chooseCountry,
+      showTrends: !uiState.showTrends,
+      currentCountry: 'USA'
+    }));
+  }
+
+  const onBackHandler = () => {
+    setUIState(prev => ({
+      ...prev,
+      chooseCountry: true,
+      showTrends: false
+    }))
+    props.deactivateContainer()
+  }
+
+  const activateTrendi = () => {
+    setUIState(prev => ({
+      ...prev,
+      chooseCountry: false,
+      showTrends: false
+    }));
+    props.activateContainer();
+  }
+
   const uiData = {
     uiState,
     toggleLeft,
     toggleRight,
     activateContainer,
-    deactivateContainer
+    deactivateContainer,
+    toggleChooseCanada,
+    toggleChooseUsa,
+    onBackHandler,
+    activateTrendi
   };
 
   return (
